test: split security URL validation tests
